src/python/web_frontend/web_requirements.txt
# Web frontend dependencies for ML Inference Offloading Dashboard
Flask==2.3.3
Flask-SocketIO==5.3.6
matplotlib>=3.5.0
pydantic>=1.8.0
PyYAML>=6.0
pyzmq>=24.0.0